refactor: consolidate records and recurrences data management, improve code organization and remove unused components
