+++ apps/mobile/src/components/finance/recurrences/ConfirmCard/index.tsx
@@ -2,7 +2,7 @@ import { Pressable, View } from 'react-native'
 import { AppIcon } from '@/components/ui/AppIcon'
 import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { useConfirmCard } from '@/hooks/useConfirmCard'
+import { useConfirmCard } from '@/hooks/recurrences/useConfirmCard'
 import { formatCurrency } from '@/utils/formatters'
 import { styles } from './style'
 import type { Recurrence } from '../types'
