+++ apps/mobile/src/components/finance/recurrences/ConfirmacoesPendentes/index.tsx
@@ -2,7 +2,7 @@ import { Animated, Pressable, View } from 'react-native'
 import { AppIcon } from '@/components/ui/AppIcon'
 import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { useConfirmacoesPendentes } from '@/hooks/useConfirmacoesPendentes'
+import { useConfirmacoesPendentes } from '@/hooks/recurrences/useConfirmacoesPendentes'
 import { ConfirmCard } from '../ConfirmCard'
 import { styles } from './style'
 import type { Recurrence } from '../types'
