+++ apps/mobile/src/components/finance/recurrences/PaymentHistoryTimeline/index.tsx
@@ -2,7 +2,7 @@ import { View } from 'react-native'
 import { AppIcon } from '@/components/ui/AppIcon'
 import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import type { PaymentHistoryEntry } from '@/components/finance/recurrences/recurrences-data'
+import type { PaymentHistoryEntry } from '@/utils/fixtures/recurrences'
 import { formatCurrency } from '@/utils/formatters'
 import { getPaymentHistoryEntryDisplay } from '@/utils/recurrences/paymentHistory'
 import { styles } from './style'
