+++ apps/mobile/src/components/finance/recurrences/RecurrenceItem/index.tsx
@@ -2,7 +2,7 @@ import { Pressable, View } from 'react-native'
 import { AppIcon } from '@/components/ui/AppIcon'
 import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { useRecurrenceItem } from '@/hooks/useRecurrenceItem'
+import { useRecurrenceItem } from '@/hooks/recurrences/useRecurrenceItem'
 import { formatCurrency } from '@/utils/formatters'
 import { styles } from './style'
 import type { Recurrence } from '../types'
