--- apps/mobile/src/components/ui/ButtonPrimary/style.ts
@@ -1,18 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export const styles = StyleSheet.create({
-  button: {
-    width: '70%',
-    height: 52,
-    borderRadius: 5,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonDisabled: {
-    opacity: 0.35,
-  },
-  text: {
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-})
