--- apps/mobile/src/components/ui/ScrollArea/style.ts
@@ -1,12 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export const styles = StyleSheet.create({
-  scrollArea: {
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-    textAlign: 'center',
-    padding: 0,
-    gap: 0,
-    width: '100%',
-  },
-})
